admin_bkend/admin/cities/serializers.py
from rest_framework import serializers

from .models import WeatherData, Cities


class WeatherDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = WeatherData
        fields = '__all__'


class CitiesSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cities
        fields = '__all__'
