admin_bkend/admin/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
mysqlclient==2.0.3
django-mysql==3.12.0
django-cors-headers==3.7.0
pika==1.1.0